update1 by dql

+++ JD-database/src/db/test/AddrDaoImplTest.java
@@ -35,7 +35,7 @@ public class AddrDaoImplTest {
 
 	@Test
 	public void testFindByID() {
-		/*System.out.println("通过主键输出一行信息");
+		System.out.println("通过地址ID输出一行信息");
 		AddrDao dao = new AddrDaoImpl();
 		int i=13;
 		Addr a = dao.findByID(i);
@@ -44,16 +44,12 @@ public class AddrDaoImplTest {
 			System.out.println("没有地址ID为 "+i+" 的地址信息");
 		}
 		else {
-			System.out.println(a.getAddr_id()+" "+a.getUser_user_id()+" "+a.getAddr_name()+" "+
-				a.getAddr_area()+" "+a.getAddr_city()+" "+a.getAddr_country()+" "+
-				a.getAddr_detail()+" "+a.getAddr_province()+" "+a.getAddr_tel()+" ");
 			System.out.println("查找地址ID为"+a.getAddr_id()+" 的一行地址信息成功!");}
-	*/
 	}
 
 	@Test
 	public void testFindAll() {
-		/*System.out.println("列出所有地址信息");
+		System.out.println("列出所有地址信息");
 		AddrDao ad = new AddrDaoImpl();
 		List<Addr> addr = ad.findAll();
 		for(int i = 0; i < addr.size(); i++) {
@@ -67,7 +63,7 @@ public class AddrDaoImplTest {
 
 	@Test
 	public void testFindByUserID() {
-		System.out.println("通过查找用户ID输出所有地址信息");
+		/*System.out.println("通过查找用户ID输出所有地址信息");
 		AddrDao ad = new AddrDaoImpl();
 		int i=126;
 		List<Addr> addr = ad.findByUserID(i);
@@ -82,48 +78,55 @@ public class AddrDaoImplTest {
 
 	@Test
 	public void testSave() {
-		System.out.println("增加地址信息");
-		int i=18;
-		AddrDao dao = new AddrDaoImpl();
-		Addr addr = new Addr();
-		addr.setAddr_id(18);
-		addr.setAddr_area("abc");
-		addr.setAddr_city("abc");
-		addr.setAddr_country("abc");
-		addr.setAddr_detail("abc");
-		addr.setAddr_is_default(true);
-		addr.setAddr_name("abc");
-		addr.setAddr_province("abc");
-		addr.setAddr_tel("123");
-		addr.setUser_user_id(126);
-		dao.save(addr);
+		/*System.out.println("增加地址信息");
+		int i=13;
 		int j=19;
-		addr.setAddr_id(j);
-		addr.setAddr_area("abc");
-		addr.setAddr_city("abc");
-		addr.setAddr_country("abc");
-		addr.setAddr_detail("abc");
-		addr.setAddr_is_default(true);
-		addr.setAddr_name("abc");
-		addr.setAddr_province("abc");
-		addr.setAddr_tel("123");
-		addr.setUser_user_id(126);
-		dao.save(addr);
-		//验证是否增加地址信息成功
-		Addr a = dao.findByID(i);
-		Addr b = dao.findByID(j);
-		if(a==null||b==null) {
-			System.out.println("Save Error!");
+		AddrDao dao = new AddrDaoImpl();
+		Addr a1=dao.findByID(i);
+		Addr a2=dao.findByID(j);
+		if(a1!=null||a2!=null) {
+			System.out.println("已经存在该地址ID!");
 		}
 		else {
-			System.out.println("增加地址ID为 "+i+" 和 "+j+" 的地址信息成功!");
-		}
+			Addr addr = new Addr();
+			addr.setAddr_id(i);
+			addr.setAddr_area("abc");
+			addr.setAddr_city("abc");
+			addr.setAddr_country("abc");
+			addr.setAddr_detail("abc");
+			addr.setAddr_is_default(true);
+			addr.setAddr_name("abc");
+			addr.setAddr_province("abc");
+			addr.setAddr_tel("123");
+			addr.setUser_user_id(126);
+			dao.save(addr);
+			addr.setAddr_id(j);
+			addr.setAddr_area("abc");
+			addr.setAddr_city("abc");
+			addr.setAddr_country("abc");
+			addr.setAddr_detail("abc");
+			addr.setAddr_is_default(true);
+			addr.setAddr_name("abc");
+			addr.setAddr_province("abc");
+			addr.setAddr_tel("123");
+			addr.setUser_user_id(126);
+			dao.save(addr);
+			//验证是否增加地址信息成功
+			Addr a = dao.findByID(i);
+			Addr b = dao.findByID(j);
+			if(a==null||b==null) {
+				System.out.println("Save Error!");
+			}
+			else {
+				System.out.println("增加地址ID为 "+i+" 和 "+j+" 的地址信息成功!");
+			}
+		}*/
 	}
 	
 	@Test
 	public void testUpdate() {
 		System.out.println("更新一行信息某列元素");
-		int i=20;
+		int i=13;
 		AddrDao dao=new AddrDaoImpl();
 		Addr addr=dao.findByID(i);
 		//判断该地址ID是否存在
@@ -145,15 +148,17 @@ public class AddrDaoImplTest {
 	
 	@Test
 	public void testDelete() {
-		/*System.out.println("删除一行指定地址ID信息");
+		System.out.println("删除一行指定地址ID信息");
 		int i=19;
 		AddrDao dao=new AddrDaoImpl();
 		Addr addr= dao.findByID(i);
-		try {
-			dao.delete(addr);
-		}catch (Exception e) {
-			//System.err.println("Delete Error!");
-			System.out.println(e.getMessage());
+		if(addr!=null) {
+			try {
+				dao.delete(addr);
+			}catch (Exception e) {
+				//System.err.println("Delete Error!");
+				System.out.println(e.getMessage());
+			}
 		}
 		//验证是否删除地址信息成功
 		Addr a = dao.findByID(i);
@@ -162,6 +167,6 @@ public class AddrDaoImplTest {
 		}
 		else {
 			System.out.println("删除地址ID为 "+i+" 的地址信息成功!");
-		}*/
+		}
 	}
 }
